Provide explicit instructions for new and old HACS views
A recent HACS update provides a new data table view of integrations. By default this only shows downloaded integrations, so leaves uses in a potentially confusing state after adding a custom repo where they cannot find the integration.  This README change provides instructions that allow overcoming this point.

We also keep and improve wording that applies for the older HACS view, as it seems to be the default for some people still.

This should be revisited in a few months, when perhaps HACS settles on a single view and the wording can be simplified. 

Resolves https://github.com/davidn/epa_uvindex/issues/5